# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API E-Commerce UTP",
        "description": "Dokumentasi API untuk UTP dengan Mock Data",
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Menampilkan semua item barang",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Membuat item baru",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Topi Baseball"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 80000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item berhasil dibuat"
                    },
                    "422": {
                        "description": "Error validasi"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Menampilkan item berdasarkan ID",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update seluruh data item",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Sepatu Nike"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 600000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seluruh data barang berhasil diedit"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    },
                    "422": {
                        "description": "Error validasi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Hapus item berdasarkan ID",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang berhasil dihapus"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Update sebagian data item",
                "operationId": "afef220a7f9ed8d384dc317e7ed7ff17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Sepatu Nike"
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 600000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data barang berhasil diedit sebagian"
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}